Removed title from BookEntity
